Changes to match the review
- Moved the styling for the OrderWords exercise (ExerciseTypeOW.sc.js)
- Reverted all changes in the Exercise.sc.js
- Moved code to smaller functions within the handleCheck action.
