+++ src/exercises/exerciseTypes/Exercise.sc.js
@@ -3,12 +3,7 @@ import { Link } from "react-router-dom";
 import {
   zeeguuTransparentLightOrange,
   zeeguuOrange,
-  darkGreen,
-  zeeguuRed,
-  zeeguuDarkOrange,
-  darkBlue,
 } from "../../components/colors";
-import { green } from "@mui/material/colors";
 
 const Exercise = styled.div`
   display: flex;
@@ -18,6 +13,7 @@ const Exercise = styled.div`
   padding-bottom: 1em;
 
   .headline {
+    font-size: ;
     color: gray;
     margin-top: 1em;
     font-weight: 500;
@@ -28,7 +24,6 @@ const Exercise = styled.div`
     color: gray;
     margin-top: 3em;
     margin-bottom: 2em;
-    
     /* font-weight: 600; */
   }
 
@@ -54,189 +49,6 @@ const Exercise = styled.div`
   .highlightedWord {
     color: orange;
     font-weight: 800;
-  }
-
-  .headlineOrderWords {
-    font-size: small;
-    color: gray;
-    margin-top: 1.5em;
-    margin-bottom: 1em;
-
-
-    /* font-weight: 600; */
-    h2 {
-      font-size: medium;
-      margin-left:8px;
-      margin-right:8px;
-    }
-  }
-
-  .cluesRow {
-    text-align: center;
-    display: inline-block;
-    margin-left: 0em;
-    margin-top: -25px;
-    
-    h4 {
-      margin-top: 0px;
-      margin-bottom: 4px;
-    }
-    p {
-      margin-bottom: 2px;
-      margin-top: 0px;
-    }
-
-  }
-
-  .orderWordsItem{
-    display: flex;
-    align-items: left;
-    justify-content: left;
-    min-height: 50px;
-    flex-wrap: wrap;
-    border-style: solid;
-    border-radius: 16px;
-    background-color: #dbdbdba6;
-    border: 1px;
-    padding: 5px 5px;
-    margin-left: 30px;
-    margin-right: 30px;
-    margin-bottom: 1em;
-    @media (max-width: 430px) {
-      font-size: 14px;
-    }
-  }
-
-
-  .incorrect {
-    background-color: ${zeeguuRed};
-    
-  }
-
-  .correct {
-    background-color: ${darkGreen};
-    :hover{ }
-  }
-
-  .swapModeBar {
-    //background-color: ${darkBlue};
-    background-color: #6db9d92b;
-    margin: 0px 30px;
-    border-top: 3px solid ${darkBlue};
-    border-bottom: 3px solid ${darkBlue};
-    border-radius: 5px;
-    margin-bottom: 1em;
-    text-align: center;
-    font-size: large;
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    min-height: 34px;
-    p {
-      text-align: center;
-      margin: 0px;
-      //color: white;
-    }
-
-    button {
-      // Compensate the Margin
-      font-size: small;
-      font-weight: bold;
-      margin-top: -2px;
-      padding: 0.5em 0.5em;
-      max-height: 28px;
-    }
-    @media (max-width: 430px) {
-      font-size: 15px;
-    }
-  }
-
-  .check {
-      background-color: darkgreen;
-      //outline: 3px darkgreen solid;
-  }
-
-  .undo {
-      background-color: ${zeeguuRed};
-
-  }
-
-  .owButton{
-    
-    cursor: pointer;
-    font-weight: 600;
-    text-align: center;
-    color: #fff;
-    border: none;
-    padding: 0.5em 0.5em;
-    border-radius: 100px;
-    box-shadow: 0px 2px #595959;
-
-    :hover {filter: brightness(0.8);}
-
-    :active {
-      filter: brightness(0.5);
-      box-shadow: 0 1px #666;
-      transform: translateY(2px);
-    }
-  }
-
-  .resetConfirmBar {
-    //background-color: ${darkBlue};
-    background-color: #ffbb545e;
-    margin: 0px 30px;
-    border-top: 3px solid ${zeeguuOrange};
-    border-bottom: 3px solid ${zeeguuOrange};
-    border-radius: 5px;
-    margin-bottom: 1em;
-    text-align: center;
-    font-size: large;
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    min-height: 34px;
-    @media (max-width: 430px) {
-      font-size: 16px;
-    }
-    p {
-      text-align: center;
-      margin: 0px;
-      //color: white;
-    }
-    button {
-      // Compensate the Margin
-      font-size: small;
-      margin-top: -2px;
-      padding: 0.5em 0.5em;
-      max-height: 28px;
-    }
-
-  }
-
-  .toSwap {
-    outline: 4px ${darkBlue} dashed;
-    background-color: #6db9d92b !important;
-    color: #000000bf;
-  }
-
-  .greyOut {
-    background-color: #595959;
-  }
-  
-  .disable {
-    background-color: #595959;
-    
-    :hover{
-      filter: brightness(1);
-    }
-    :active {
-      filter: brightness(1);
-      box-shadow: 0px 2px #595959;
-      transform: translateY(0px);
-    }
-  }
-  .tipText{
-    font-size:small
   }
 
   .contextExample {
@@ -342,71 +154,6 @@ let OrangeButton = styled(StyledButton)`
   color: white;
   background-color: ${zeeguuOrange};
   margin: 1em;
-`;
-
-let OrangeItemCompact = styled(StyledButton)`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  font-size: normal !important;
-  color: white;
-  height: 34px;
-  background-color: ${zeeguuOrange};
-  margin: 0.2em 0.2em;
-  padding: 0.5em 0.5em;
-  box-shadow: 0 1px #666;
-  &:hover {
-    filter: brightness(0.8);
-  }
-  &:active {
-      filter: brightness(0.5);
-      box-shadow: 0 0.5px #666;
-      transform: translateY(1px);
-    }
-  
-`;
-
-let OrangeItemCompactConstruct = styled(StyledButton)`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  font-size: normal !important;
-  align-items: center;
-  color: white;
-  height: 30px;
-  background-color: ${zeeguuOrange};
-  margin: 0.05em;
-  padding: 0.5em 0.5em;
-  & .correct :hover {
-  }
-
-  & .toSwap {
-    color:black;
-  }
-  @media (max-width: 430px) {
-    font-size: 10px;
-  }
-  /*
-  & .feedbackText {
-    visibility: hidden;
-    width: 150px;
-    overflow: visible;
-    background-color: black;
-    color: #fff;
-    text-align: center;
-    border-radius: 6px;
-    padding: 5px;
-    /* Position the tooltip 
-    position: absolute;
-    z-index: 1;
-    margin-bottom: 110px;
-  }
-
-  &:hover .feedbackText {
-    visibility: visible;
-  }
-  */
 `;
 
 let OrangeButtonMessage = styled(StyledButton)`
@@ -535,30 +282,6 @@ let BottomRowCompact = styled.div`
   }
 `;
 
-let ItemRowCompactWrap = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  padding: 0.5em;
-  align-items: center;
-  justify-content: space-around;
-  margin-bottom: 1em;
-  @media (max-width: 430px) {
-    flex-flow: row wrap;
-  }
-`;
-let ItemRowCompactWrapConstruct = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  padding: 0.3em;
-  align-items: left;
-  min-height:64px;
-  //justify-content: ;
-  margin-bottom: 0.1em;
-  @media (max-width: 430px) {
-    flex-flow: row wrap;
-  }
-`;
-
 let StyledGreyButton = styled.button`
   background: none !important;
   border: none !important;
@@ -609,15 +332,11 @@ export {
   Exercise,
   FeedbackButton,
   OrangeButton,
-  OrangeItemCompact,
-  OrangeItemCompactConstruct,
   AnimatedOrangeButton,
   Input,
   AnimatedInput,
   BottomRow,
   BottomRowCompact,
-  ItemRowCompactWrap,
-  ItemRowCompactWrapConstruct,
   StyledGreyButton,
   StyledDiv,
   MatchButton,
